+++ client/src/pages/StudentDashboard.js
@@ -5,10 +5,28 @@ import WelcomeMsg from "../components/WelcomeMsg";
 import Profile from "../components/Profile";
 import FeedbackModal from "../components/FeedbackModal";
 import Progress from "../components/Progress";
+import { Upload } from "upload-js";
 
 const StudentDashboard = ({ theme }) => {
 	const [user, setUser] = React.useState({});
+	const [bio, setBio] = React.useState(user.bio);
+	const [techModule, setTechModule] = React.useState({
+		module: "",
+		lesson: "",
+	});
+	const [sofSkill, setSoftSkill] = React.useState("");
+	const [avatarUrl, setAvatarUrl] = React.useState("");
+	const upload = Upload({ apiKey: "free" });
 
+	async function onFileSelected(event) {
+		const [file] = event.target.files;
+		const { fileUrl } = await upload.uploadFile(file, {
+			onBegin: ({ cancel }) => console.log("File upload started!"),
+			onProgress: ({ progress }) =>
+				console.log(`File uploading... ${progress}%`),
+		});
+		setAvatarUrl(fileUrl);
+	}
 	const getUserById = async () => {
 		try {
 			const user = JSON.parse(localStorage.getItem("user"));
@@ -29,16 +47,6 @@ const StudentDashboard = ({ theme }) => {
 		getUserById();
 	}, []);
 
-	const [bio, setBio] = React.useState(
-		"Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec nec consequat ipsum."
-	);
-	const [techModule, setTechModule] = React.useState({
-		module: "",
-		lesson: "",
-	});
-
-	const [sofSkill, setSoftSkill] = React.useState("");
-
 	// handle input change functions
 	const handleBioChange = (e) => {
 		setBio(e.target.value);
@@ -52,6 +60,7 @@ const StudentDashboard = ({ theme }) => {
 	function handleSoftSkillChange(e, newValue) {
 		setSoftSkill(newValue);
 	}
+
 	return (
 		<ThemeProvider theme={theme}>
 			<Stack
@@ -65,8 +74,10 @@ const StudentDashboard = ({ theme }) => {
 				<WelcomeMsg message={`Welcome ${user.full_name}!👋`} />
 				<Profile
 					mentorName="mentor name"
+					avatar={user.img_url || avatarUrl}
 					bio={bio}
 					handleBioChange={handleBioChange}
+					onFileSelected={onFileSelected}
 				/>
 				<Progress
 					techModule={techModule}
